backend/pyproject.toml
[project]
name = "agente"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai",
    "browser-use",
    "python-dotenv",
    "fastapi",
    "fastapi[standard]",
    "uvicorn",
    "asyncio",
    "pydantic",
    "pydantic-settings",
    "pytz",
    "apscheduler",
    "requests",
    "supabase",
    "pydantic",
    "pyjwt",
    "fastapi-jwt-auth",
]
